% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_preprocessing.R
\name{log_transform}
\alias{log_transform}
\title{Log-transform a strictly positive column into \verb{<col>_log}}
\usage{
log_transform(df, col)
}
\arguments{
\item{df}{Portfolio tibble.}

\item{col}{Column name.}
}
\value{
\code{df} with an added column \verb{<col>_log}.
}
\description{
Log-transform a strictly positive column into \verb{<col>_log}
}
